Move marker trait and add docs
